+++ meridian/analysis/optimizer.py
@@ -936,17 +936,23 @@ class BudgetOptimizer:
         used in conjunction to determine the non-optimized media-level spend,
         which is used to calculate the non-optimized performance metrics (for
         example, ROI) and construct the feasible range of media-level spend with
-        the spend constraints.
+        the spend constraints. Consider using
+        `InputData.get_paid_channels_argument_builder()` to construct the
+        argument here.
       spend_constraint_lower: Numeric list of size `n_total_channels` or float
         (same constraint for all channels) indicating the lower bound of
         media-level spend. The lower bound of media-level spend is `(1 -
         spend_constraint_lower) * budget * allocation)`. The value must be
         between 0-1. Defaults to `0.3` for fixed budget and `1` for flexible.
+        Consider using `InputData.get_paid_channels_argument_builder()` to
+        construct the argument here.
       spend_constraint_upper: Numeric list of size `n_total_channels` or float
         (same constraint for all channels) indicating the upper bound of
         media-level spend. The upper bound of media-level spend is `(1 +
         spend_constraint_upper) * budget * allocation)`. Defaults to `0.3` for
-        fixed budget and `1` for flexible.
+        fixed budget and `1` for flexible. Consider using
+        `InputData.get_paid_channels_argument_builder()` to construct the
+        argument here.
       target_roi: Float indicating the target ROI constraint. Only used for
         flexible budget scenarios. The budget is constrained to when the ROI of
         the total spend hits `target_roi`.