+++ meridian/data/__init__.py
@@ -14,6 +14,7 @@
 
 """Data handling API for Meridian."""
 
+from meridian.data import arg_builder
 from meridian.data import input_data
 from meridian.data import load
 from meridian.data import time_coordinates